requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-motor==3.4.14
adafruit-circuitpython-pca9685==3.4.16
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-servokit==1.3.18
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
attrs==24.2.0
binho-host-adapter==0.1.6
contourpy==1.3.1
cycler==0.12.1
evdev==1.7.1
fonttools==4.55.0
kiwisolver==1.4.7
numpy==1.26.4
packaging==24.2
pillow==11.0.0
pyarrow==18.0.0
pyftdi==0.55.4
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-xlib==0.33
pyusb==1.2.1
rerun-sdk==0.20.1
ruff==0.7.3
six==1.16.0
sysv_ipc==1.1.0
typing_extensions==4.12.2
